requirements.txt
requests~=2.31.0
python-socketio==5.11.2
websocket-client==1.8.0
aiohttp==3.9.5
langdetect==1.0.9
